move notifications button to _navigation directory
