+++ packages/atlas/src/components/_navigation/TopbarStudio/TopbarStudio.tsx
@@ -3,9 +3,9 @@ import { CSSTransition } from 'react-transition-group'
 
 import { AvatarGroupUrlAvatar } from '@/components/Avatar/AvatarGroup'
 import { Button } from '@/components/_buttons/Button'
-import { NotificationsButton } from '@/components/_buttons/NotificationsButton'
 import { SvgActionAddVideo } from '@/components/_icons'
 import { SvgJoystreamLogoStudio } from '@/components/_illustrations'
+import { NotificationsButton } from '@/components/_navigation/NotificationsButton'
 import { MemberDropdown } from '@/components/_overlays/MemberDropdown'
 import { absoluteRoutes } from '@/config/routes'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
