+++ packages/atlas/src/components/_navigation/TopbarViewer/TopbarViewer.tsx
@@ -4,10 +4,10 @@ import { CSSTransition, SwitchTransition } from 'react-transition-group'
 
 import { Searchbar } from '@/components/Searchbar'
 import { Button } from '@/components/_buttons/Button'
-import { NotificationsButton } from '@/components/_buttons/NotificationsButton'
 import { SvgActionMember } from '@/components/_icons'
 import { SvgJoystreamLogoFull } from '@/components/_illustrations'
 import { SkeletonLoader } from '@/components/_loaders/SkeletonLoader'
+import { NotificationsButton } from '@/components/_navigation/NotificationsButton'
 import { MemberDropdown } from '@/components/_overlays/MemberDropdown'
 import { QUERY_PARAMS, absoluteRoutes } from '@/config/routes'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
